`viking`: Fix usage of data_symbols
When trying to resolve data symbol-pointers, the checker tries to account the file offset of the pointer twice: Once in the check_data_symbol_ptr itself, and once again when calling elf::get_bytes. This lead to the pointers pointing to a wrong offset, resolving into different data structures and thus displaying (nonexistant) errors or just failing to resolve due to an out-of-bounds read.

This PR fixes this behaviour, by removing the manual offset calculation from this class and just letting that handle the elf class.