viking: streamline new function diffing process
Instead of running print_decomp_symbols.py or listsym <function name>
and then pasting the mangled name in the CSV, users can now just paste
the human-readable function name (e.g. Ecosystem::init) into the CSV;
the checker will attempt to find the correct mangled name, update
the CSV and use that for diffing.

If there are several matches an interactive prompt is shown so that
the user can select the correct function to diff.